requirements.txt
xlwings==0.30.2
tenacity==8.2.3
psutil==5.9.8
office365-rest-python-client==2.5.6
python-dotenv==1.0.1
